refactor: what an app is doing is read the same way on both surfaces
The status model was the dashboard's own, and the CLI decides the same
question — is anything running, has anything ever been deployed — off
`app.state` and a release state at each call site.

It moves to `@repo/app-operations` unchanged, so both surfaces name the
eleven states the same way.
