+++ apps/dashboard/src/lib/app-actions.ts
@@ -1,4 +1,4 @@
-import { type AppStatus, type AppStatusKey, statusKey } from '#lib/app-status.ts';
+import { type AppStatus, type AppStatusKey, statusKey } from '@repo/app-operations';
 
 /** Every button on an app's action bar, and so every column of the table below. */
 export const APP_ACTIONS = ['deploy', 'export', 'suspend', 'delete'] as const;
